Refactor library export artifact emitters
